// api/test/mocks.ts
export enum ImageFiles {
  JPG_Example_1 = 'test-jpg-image-1.jpeg',
  JPG_Example_2 = 'test-jpg-image-2.jpeg',
  JPG_Example_3 = 'test-jpg-image-3.jpeg',
  JPG_Example_4 = 'test-jpg-image-4.jpeg',
  JPG_Example_5 = 'test-jpg-image-5.jpeg',
  PNG_Example = 'test-png-image.png',
  SVG_Example = 'test-svg-image.svg',
}